refactor: :recycle: refactoring ChapterVersesPage
